// nest-10/src/service-now-insights-widget-codes.enum.ts
export enum WidgetCodes {
    CATEGORY_DISTRIBUTION = "d757b6f7-50e4-4aa5-9480-419ab5fc94e2",
    CRITICAL_BACKLOG = "e613ba75-c4f1-4064-b5c7-bcba87066b25",
    GROUP_WORKLOAD = "a9148959-3af4-4233-a7c0-879b58d656cc",
    SLA_BREACH_STATUS = "704f792f-fa42-4090-aa80-ea29f73945cf",
    STALE_TICKETS = "ab8fdd23-9d3a-4512-8140-6e2e358ad329",
    STATE_FUNNEL = "be973db9-a82d-466e-9b75-d6f54780a1f7",
    TOP_MONTHLY_CALLERS = "514d747d-a89f-4e19-945c-05a78e4152a2",
    TOTAL_OPEN_INCIDENTS = "9a2351c6-f749-4743-b091-aab16fd1e1d3",
    UNASSIGNED_TICKETS = "d905b33d-46d5-498d-af86-e01a2e775908",
    VOLUME_BY_PRIORITY = "a11e181a-7e6c-43f2-9150-a616bec962c9"
}